+++ matsim/src/main/java/org/matsim/utils/gis/shp2matsim/Shp2NetworkModification.java
@@ -17,55 +17,47 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.contrib.drt.util;
+package org.matsim.utils.gis.shp2matsim;
 
-import java.io.File;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.Collection;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
+import org.apache.log4j.Logger;
 import org.locationtech.jts.geom.Geometry;
-import org.locationtech.jts.geom.Point;
-import org.matsim.api.core.v01.Coord;
-import org.matsim.core.utils.geometry.geotools.MGC;
-import org.matsim.core.utils.gis.ShapeFileReader;
-import org.opengis.feature.simple.SimpleFeature;
+import org.matsim.api.core.v01.Scenario;
+import org.matsim.api.core.v01.TransportMode;
+import org.matsim.api.core.v01.network.Link;
 
-public class DrtShpUtils {
+public class Shp2NetworkModification {
 	
-	public static List<Geometry> loadShapeFile(String shapeFile) {
-		List<Geometry> geometries = new ArrayList<>();
+	private static final Logger LOGGER = Logger.getLogger( Shp2NetworkModification.class );
 
-		Collection<SimpleFeature> features = null;
-		if (new File(shapeFile).exists()) {
-			features = ShapeFileReader.getAllFeatures(shapeFile);	
-		} else {
-			try {
-				features = ShapeFileReader.getAllFeatures(new URL(shapeFile));
-			} catch (MalformedURLException e) {
-				e.printStackTrace();
+	public static void addModeToLinksStartingOrEndingInShp(Scenario scenario, String modeToAdd, String shpFile) {
+		
+		LOGGER.info("Adding drt mode to network...");
+		
+		List<Geometry> serviceAreaGeometries = ShpGeometryUtils.loadShapeFile(shpFile);
+
+		int counter = 0;
+		for (Link link : scenario.getNetwork().getLinks().values()) {
+			if (counter % 10000 == 0)
+				LOGGER.info("link #" + counter);
+			counter++;
+			if (link.getAllowedModes().contains(TransportMode.car)) {
+				if ( ShpGeometryUtils.isCoordInGeometries(link.getFromNode().getCoord(), serviceAreaGeometries)
+						|| ShpGeometryUtils.isCoordInGeometries(link.getToNode().getCoord(), serviceAreaGeometries) ) {
+					
+					Set<String> allowedModes = new HashSet<>(link.getAllowedModes());
+					allowedModes.add(modeToAdd);
+					link.setAllowedModes(allowedModes);
+				}
+
+			} else if (link.getAllowedModes().contains(TransportMode.pt)) {
+				// skip pt links
+			} else {
+				throw new RuntimeException("Aborting...");
 			}
 		}
-		
-		if (features == null) throw new RuntimeException("Aborting...");
-		
-		for (SimpleFeature feature : features) {
-			geometries.add( (Geometry) feature.getDefaultGeometry() );
-		}
-		return geometries;
 	}
-
-	public static boolean isCoordInGeometries( Coord coord, List<Geometry> geometries ) {
-		Point p = MGC.coord2Point(coord);
-		
-		for (Geometry geometry : geometries) {
-			if (p.within(geometry)) {
-				return true;
-			}
-		}
-		return false;
-	}
-
 }
